Split out PDF context manager
